refactor commits - clean now
