+++ src/manipulation/scripts/visual_servoing.py
@@ -19,7 +19,6 @@ from cv_bridge import CvBridge, CvBridgeError
 from helpers import convertPointToDepth
 import cv2
 from scipy.spatial.transform import Rotation as R
-from geometry_msgs.msg import Vector3, Vector3Stamped, PoseStamped, Point, Pose, PointStamped
 from sensor_msgs.msg import PointCloud2, PointField
 import pandas as pd
 # ROS action finite state machine
@@ -27,8 +26,7 @@ from control_msgs.msg import FollowJointTrajectoryAction
 from control_msgs.msg import FollowJointTrajectoryResult
 import actionlib
 from yolo.msg import Detections
-from termcolor import colored
-import message_filters
+from scene_parser import SceneParser
 
 class State(Enum):
     SEARCH = 1
@@ -44,188 +42,52 @@ class State(Enum):
 
     FAILED = -1
 
-class Error(Enum):
-    CLEAN = 0
-    SCAN_FAILED = 1
-    
 
 class AlignToObject:
-    def __init__(self, objectId):
+    def __init__(self, objectId, scene_parser : SceneParser):
         self.objectId = objectId
         self.node_name = 'visual_alignment'
         self.state = State.SEARCH
-        self.rate = 10.0
-        self.kp = {
-            'velx' : 0.7
-        }
-        self.kd =   {
-            'velx' : 0.0
-        }         
-
+        
+        self.scene_parser = scene_parser
+        self.pid_gains = rospy.get_param('/manipulation/pid_gains')
+        
         self.prevxerr = 0    
-        self.isDetected = False 
         
-        self.vs_range = [(-np.deg2rad(60), np.deg2rad(60)), 12] # [(left angle, right angle), stepsize]
-
-        self.trajectoryClient = actionlib.SimpleActionClient('alfred_controller/follow_joint_trajectory', FollowJointTrajectoryAction)
-        self.listener = tf.TransformListener()
-        self.tf_ros = TransformerROS()
-
-
-        rospy.loginfo(f"[{rospy.get_name()}]: Waiting for camera intrinsics...")
-        msg = rospy.wait_for_message('/camera/color/camera_info', CameraInfo, timeout=10)
-        self.intrinsics = np.array(msg.K).reshape(3,3)
-        self.cameraParams = {
-            'width' : msg.width,
-            'height' : msg.height,
-            'intrinsics' : np.array(msg.K).reshape(3,3),
-            'focal_length' : self.intrinsics[0][0],
-        }
-        rospy.loginfo(f"[{rospy.get_name()}]: Obtained camera intrinsics.")
-
+        self.recoveries = rospy.get_param('/manipulation/recoveries')
+        self.mean_err_horiz_alignment = rospy.get_param('/manipulation/mean_err_horiz_alignment')
+        self.vs_range = rospy.get_param('/manipulation/vs_range') 
+        self.vs_range[0] *= np.pi/180
+        self.vs_range[1] *= np.pi/180
         
-        self.bridge = CvBridge()
+        self.moving_avg_n = rospy.get_param('/manipulation/moving_avg_n')
+        self.max_graspable_distance = rospy.get_param('/manipulation/max_graspable_distance')
         
-        self.obtainedInitialImages = False
-
-        self.maxDistanceToMove = 0.0
-
-        self.objectLocationInCamera = [np.inf, np.inf, np.inf]
-
-        self.requestClearObject = False
-        self.objectLocArr = []
-        self.upscale_fac = 1/rospy.get_param('/image_shrink/downscale_ratio')
+        self.head_tilt_angle_search = rospy.get_param('/manipulation/head_tilt_angle_search')
+        self.head_tilt_angle_grasp = rospy.get_param('/manipulation/head_tilt_angle_grasp')
+        
+        
+        self.forward_vel = rospy.get_param('/manipulation/forward_velocity')
+        self.interval_before_restart = rospy.get_param('/manipulation/forward_velocity')
         
         self.obj_tf_publisher = tf.TransformBroadcaster()
-        self.depth_image_subscriber = message_filters.Subscriber('/camera/aligned_depth_to_color/image_raw', Image)
-        self.rgb_image_subscriber = message_filters.Subscriber('/camera/color/image_raw', Image)
-        self.boundingBoxSub = message_filters.Subscriber("/object_bounding_boxes", Detections)
         
-        synced_messages = message_filters.ApproximateTimeSynchronizer([self.depth_image_subscriber, self.rgb_image_subscriber, self.boundingBoxSub], 60, 0.5, allow_headerless=False)
-        synced_messages.registerCallback(self.parseScene)
-        rospy.loginfo("Waiting for synchronized images")
-        while not self.obtainedInitialImages:
-            rospy.sleep(0.1)
+        rospy.loginfo(f"[{self.node_name}]: Waiting for trajectory server")
+        self.trajectoryClient = actionlib.SimpleActionClient('alfred_controller/follow_joint_trajectory', FollowJointTrajectoryAction)
+        self.trajectoryClient.wait_for_server()
+        
+        
         rospy.loginfo(f"[{self.node_name}]: Initial images obtained.")
         rospy.loginfo(f"[{self.node_name}]: Node initialized")
         self.prevtime = time.time()
         
-    def create_point_cloud_from_depth_image(self, depth, organized=True):
-        """ Generate point cloud using depth image only.
-
-            Input:
-                depth: [numpy.ndarray, (H,W), numpy.float32]
-                    depth image
-                camera: [CameraInfo]
-                    camera intrinsics
-                organized: bool
-                    whether to keep the cloud in image shape (H,W,3)
-
-            Output:
-                cloud: [numpy.ndarray, (H,W,3)/(H*W,3), numpy.float32]
-                    generated cloud, (H,W,3) for organized=True, (H*W,3) for organized=False
-        """
-        camera = self.cameraParams
-        assert(depth.shape[1] == camera['width'] and depth.shape[0] == camera['height'])
-        # xmap = np.arange(-camera.width//2, camera.width//2)
-        xmap = np.arange(0, camera['width'])
-        ymap = np.arange(0, camera['height'])
-        xmap, ymap = np.meshgrid(xmap, ymap)
-        points_z = depth / 1000.0
-        self.intrinsics = camera['intrinsics']
-        
-        cx = self.intrinsics[0][2]
-        cy = self.intrinsics[1][2]
-        fx = self.intrinsics[0][0]
-        fy = self.intrinsics[1][1]
-        
-        points_x = (xmap - cx) * points_z / fx
-        points_y = (ymap - cy) * points_z / fy
-        cloud = np.stack([points_x, points_y, points_z], axis=-1)
-        if not organized:
-            cloud = cloud.reshape([-1, 3])
-        return cloud
     
-    def transform_point_cloud(self, cloud, transform, format='4x4'):
-        """ Transform points to new coordinates with transformation matrix.
-
-            Input:
-                cloud: [np.ndarray, (N,3), np.float32]
-                    points in original coordinates
-                transform: [np.ndarray, (3,3)/(3,4)/(4,4), np.float32]
-                    transformation matrix, could be rotation only or rotation+translation
-                format: [string, '3x3'/'3x4'/'4x4']
-                    the shape of transformation matrix
-                    '3x3' --> rotation matrix
-                    '3x4'/'4x4' --> rotation matrix + translation matrix
-
-            Output:
-                cloud_transformed: [np.ndarray, (N,3), np.float32]
-                    points in new coordinates
-        """
-        if not (format == '3x3' or format == '4x4' or format == '3x4'):
-            raise ValueError('Unknown transformation format, only support \'3x3\' or \'4x4\' or \'3x4\'.')
-        if format == '3x3':
-            cloud_transformed = np.dot(transform, cloud.T).T
-        elif format == '4x4' or format == '3x4':
-            ones = np.ones(cloud.shape[0])[:, np.newaxis]
-            cloud_ = np.concatenate([cloud, ones], axis=1)
-            cloud_transformed = np.dot(transform, cloud_.T).T
-            cloud_transformed = cloud_transformed[:, :3]
-        return cloud_transformed
-    
-    def get_transform(self, base_frame, camera_frame):
-        listener = self.listener
-        while not rospy.is_shutdown():
-            try:
-                t = listener.getLatestCommonTime(base_frame, camera_frame)
-                (trans,rot) = listener.lookupTransform(base_frame, camera_frame, t)
-
-                trans_mat = np.array(trans).reshape(3, 1)
-                rot_mat = R.from_quat(rot).as_matrix()
-                transform_mat = np.hstack((rot_mat, trans_mat))
-                # print(listener.allFramesAsString())
-
-                return transform_mat
-                
-            except (tf.LookupException, tf.ConnectivityException, tf.ExtrapolationException):
-                print("Not found transform sorry.")
-                continue
-    def get_grasp(self):
-        if not self.clearAndWaitForNewObject(5):
-            return [], False
-        
-        dimg = self.depth_image.copy()
-        dimg *= self.ws_mask
-        
-        cloud = self.create_point_cloud_from_depth_image(dimg, organized = False)
-        transform = self.get_transform('camera_color_optical_frame', 'base_link')
-        cloud = self.transform_point_cloud(cloud, transform, format='3x4')
-        print(cloud.shape)
-        x, y, z = np.median(cloud, axis = 0)
-        
-        return [x, y, z], True
-        
-
-    def computeObjectLocation(self, objectLocs):
-        objectLocs = np.array(objectLocs)
-        if len(objectLocs) == 0:
-            return np.inf, np.inf, np.inf, np.inf, np.inf
-
-        weights = objectLocs[:, 3]
-        x = np.average(objectLocs[:, 0], weights = weights)
-        y = np.average(objectLocs[:, 1], weights = weights)
-        z = np.average(objectLocs[:, 2], weights = weights)
-        angleToGo = np.arctan2(y, x)
-        radius = np.sqrt(x**2 + y**2)
-        return angleToGo, x, y, z, radius
-
     def findAndOrientToObject(self):
         
         rospy.loginfo(f"[{self.node_name}]: Finding and orienting to object")
         
         move_to_pose(self.trajectoryClient, {
-            'head_tilt;to' : -10 * np.pi/180,
+            'head_tilt;to' : - self.head_tilt_angle_search * np.pi/180,
             'head_pan;to' : self.vs_range[0][0],
         })
         
@@ -235,25 +97,17 @@ class AlignToObject:
                 move_to_pose(self.trajectoryClient, {
                     'head_pan;to' : angle,
                 })
-                rospy.sleep(0.5)
+                rospy.sleep(0.5) #settling time.
 
-            objectLocs = np.array(self.objectLocArr)
-            if len(objectLocs) != 0:
-                # compute some statistical measure of where the object is to make some more intelligent servoing guesses.
-                angleToGo, x, y, z, radius,  = self.computeObjectLocation(self.objectLocArr)
-
+            [angleToGo, x, y, z, radius, _], success = self.scene_parser.estimate_object_location()
+            self.scene_parser.clear_observations()
+            if success:
                 self.objectLocation = [x, y, z]
-                self.requestClearObject = True
-                # self.maxDistanceToMove = radius
-
                 rospy.loginfo('Object found at angle: {}. Location = {}. Radius = {}'.format(angleToGo, (x, y, z), radius))
-
                 move_to_pose(self.trajectoryClient, {
                     'base_rotate;by' : angleToGo,
                     'head_pan;to' : 0,
                 })
-                
-                
                 break
             else:
                 rospy.loginfo("Object not found, rotating base.")
@@ -263,68 +117,41 @@ class AlignToObject:
                 })
                 rospy.sleep(5)
                 nRotates += 1
-
-            if nRotates >= 3:
+                
+            if nRotates >= self.recoveries['n_scan_attempts']:
                 return False
-        print("Found object at ", angle)
-        rospy.sleep(1)
+        
         return True
 
     def recoverFromFailure(self):
         rospy.loginfo("Attempting to recover from failure")        
         move_to_pose(self.trajectoryClient, {
-            'base_translate;vel' : -0.1,
+            'base_translate;vel' : -self.forward_vel,
         })
         rospy.sleep(1)
         move_to_pose(self.trajectoryClient, {
             'base_translate;vel' : 0.0,
         })
-        
-    def clearAndWaitForNewObject(self, numberOfCopies = 10):
-        rospy.loginfo("Clearing object location array")
-        while self.requestClearObject:
-            rospy.sleep(0.1)
-        rospy.loginfo("Object location array cleared. Waiting for new object location")
-        startTime = time.time()
-        while len(self.objectLocArr) <= numberOfCopies:
-            rospy.loginfo(f"Accumulated copies = {len(self.objectLocArr)} ")
-            if time.time() - startTime >= 10:
-                rospy.loginfo("Object not found. Restarting search")
-                # self.recoverFromFailure()
-                # self.findAndOrientToObject()
-                return False
-            rospy.sleep(1)
 
-        return True
 
     def moveTowardsObject(self):
         rospy.loginfo("Moving towards the object")
         lastDetectedTime = time.time()
         startTime = time.time()
-        vel = 0.1
-        intervalBeforeRestart = 10
+        vel = self.forward_vel
+        intervalBeforeRestart = self.interval_before_restart
 
         move_to_pose(self.trajectoryClient, {
-            'head_tilt;to' : -30 * np.pi/180,
+            'head_tilt;to' : - self.head_tilt_angle_grasp * np.pi/180,
         })
+       
+        if not self.scene_parser.clearAndWaitForNewObject():
+            return False
 
-        if self.objectId == 60:
-            self.isDepthMatters = True
-            if not self.clearAndWaitForNewObject(2):
-                return False
+        maxGraspableDistance = self.max_graspable_distance
 
-            self.isDepthMatters = False
-        else:
-            if not self.clearAndWaitForNewObject():
-                return False
-
-        maxGraspableDistance = 0.77
-        if self.objectId == 41:
-            maxGraspableDistance = 1
-            
-        distanceToMove = self.computeObjectLocation(self.objectLocArr)[3] - maxGraspableDistance
-        rospy.loginfo("Object distance = {}".format(self.computeObjectLocation(self.objectLocArr)[4]))
-        # rospy.loginfo("Distance to move = {}. Threshold = {}".format(distanceToMove, distanceThreshold))
+        distanceToMove = self.scene_parser.estimate_object_location(self.scene_parser.objectLocArr)[3] - maxGraspableDistance
+        rospy.loginfo("Object distance = {}".format(self.scene_parser.estimate_object_location(self.scene_parser.objectLocArr)[4]))
         if distanceToMove < 0:
             rospy.loginfo("Object is close enough. Not moving.")
             return True
@@ -336,26 +163,22 @@ class AlignToObject:
             move_to_pose(self.trajectoryClient, {
                 'base_translate;vel' : vel,
             })
-            # rospy.loginfo("Distance to move = {}, Distance moved = {}".format(distanceToMove, distanceMoved))
-            # rospy.loginfo()
-            rospy.loginfo("Object location = {}".format(self.objectLocArr[-1]))
-            if self.objectId != 60:
-                if self.isDetected:
-                    lastDetectedTime = time.time()
-                    if self.objectLocArr[-1][0] < maxGraspableDistance:
-                        rospy.loginfo("Object is close enough. Stopping.")
-                        rospy.loginfo("Object location = {}".format(self.objectLocArr[-1]))
-                        break
-                if distanceMoved >= distanceToMove - 0.3:
-                    if time.time() - lastDetectedTime > intervalBeforeRestart:
-                        rospy.loginfo("Lost track of the object. Going back to search again.")
-                        move_to_pose(self.trajectoryClient, {
-                            'base_translate;vel' : 0.0,
-                        })  
-                        return False
-            else: # open loop movement for the table
-                if distanceMoved >= distanceToMove:
+            rospy.loginfo("Distance to move = {}, Distance moved = {}".format(distanceToMove, distanceMoved))
+
+            lastDetectedTime = time.time()
+            object_location, isLive = self.scene_parser.get_latest_observation()
+            if isLive:
+                if object_location[0] < maxGraspableDistance:
+                    rospy.loginfo("Object is close enough. Stopping.")
+                    rospy.loginfo("Object location = {}".format(self.scene_parser.objectLocArr[-1]))
                     break
+            if distanceMoved >= distanceToMove - 0.3:
+                if time.time() - lastDetectedTime > intervalBeforeRestart:
+                    rospy.loginfo("Lost track of the object. Going back to search again.")
+                    move_to_pose(self.trajectoryClient, {
+                        'base_translate;vel' : 0.0,
+                    })  
+                    return False
 
             rospy.sleep(0.1)
 
@@ -364,250 +187,78 @@ class AlignToObject:
         })          
         return True
 
-
-
     def alignObjectForManipulation(self):
-        print("Aligning object for manipulation")
+        """
+        Rotates base to align manipulator for grasping
+        """
+        rospy.loginfo(f"[{self.node_name}]: Aligning object for manipulation")
         move_to_pose(self.trajectoryClient, {
                 'head_pan;to' : -np.pi/2,
-            }
-        )
-        move_to_pose(self.trajectoryClient, {
                 'base_rotate;by' : np.pi/2,
             }
         )
         rospy.sleep(4)
         return True
-    
-
-    def convertPointToFrame(self, x_f, y_f, z, to_frame = "base_link"):
-        intrinsic = self.intrinsics
-        fx = intrinsic[0][0]
-        fy = intrinsic[1][1] 
-        cx = intrinsic[0][2] 
-        cy = intrinsic[1][2]
-
-        x_gb = (x_f - cx) * z/ fx
-        y_gb = (y_f - cy) * z/ fy
-
-        camera_point = PointStamped()
-        camera_point.header.frame_id = '/camera_color_optical_frame'
-        camera_point.point.x = x_gb
-        camera_point.point.y = y_gb
-        camera_point.point.z = z
-        point = self.listener.transformPoint(to_frame, camera_point).point
-        return point
-
-    
-
-    def parseScene(self, depth_img, color_img, msg):
-        depth_img = self.bridge.imgmsg_to_cv2(depth_img, desired_encoding="passthrough")
-        self.depth_image = np.array(depth_img, dtype=np.float32)
-        
-        color_img = self.bridge.imgmsg_to_cv2(color_img, desired_encoding="passthrough")
-        self.color_image = np.array(color_img, dtype=np.uint8)
-        
-        
-        if self.objectId is None:
-            return
-        self.obtainedInitialImages = True
-        # to ensure thread safe behavior, writing happens only in this function on a separate thread
-        if self.requestClearObject:
-            self.requestClearObject = False
-            self.objectLocArr = []
-
-        num_detections = (msg.nPredictions)
-        msg = {
-            "boxes" : np.array(msg.box_bounding_boxes).reshape(num_detections, 4),
-            "box_classes" : np.array(msg.box_classes).reshape(num_detections),
-            'confidences' : np.array(msg.confidences).reshape(num_detections),
-        }
-
-        loc = np.where(np.array(msg['box_classes']).astype(np.uint8) == self.objectId)[0]
-        upscale_fac = self.upscale_fac
-        
-        if len(loc) > 0:
-            loc = loc[0]
-            box = np.squeeze(msg['boxes'][loc]).astype(int)
-            confidence = np.squeeze(msg['confidences'][loc])
-            x1, y1, x2, y2 =  box
-            # print(box)
-            crop = self.depth_image[y1 : y2, x1 : x2]
-            # crop = self.depth_image[int(y1 * upscale_fac) : int(y2 * upscale_fac), int(x1 * upscale_fac) : int(x2 * upscale_fac)]
-            
-            self.ws_mask = np.zeros_like(self.depth_image)
-            self.ws_mask[y1 : y2, x1 : x2] = 1
-            
-            
-            z_f = np.median(crop[crop != 0])/1000.0
-            h, w = self.depth_image.shape[:2]
-            # y_new1 = w - x2 * upscale_fac
-            # x_new1 = y1 * upscale_fac
-            # y_new2 = w - x1 * upscale_fac
-            # x_new2 = y2 * upscale_fac
-            x_f = (x1 + x2)/2
-            y_f = (y1 + y2)/2
-            z_f = self.depth_image[int(y_f), int(x_f)]/1000.0
-            to_frame = 'base_link'
-            base_pt = self.convertPointToFrame(x_f, y_f, z_f, to_frame)
-            # print(base_pt.x, base_pt.y, base_pt.z)
-            self.obj_tf_publisher.sendTransform((base_pt.x, base_pt.y, base_pt.z),
-                    tf.transformations.quaternion_from_euler(0, 0, 0),
-                    rospy.Time.now(),
-                    "object_pose",
-                    to_frame)
-            
-            if time.time() - self.prevtime > 1:
-                viz = self.color_image.copy().astype(np.float32)
-                viz /= np.max(viz)
-                viz *= 255
-                viz = viz.astype(np.uint8)
-                # viz = cv2.cvtColor(viz, cv2.COLOR_GRAY2BGR)
-                
-                
-                
-                cv2.rectangle(viz, (x1, y1), (x2, y2), (0, 255, 0), 2)
-                cv2.circle(viz, (int(x_f), int(y_f)), 5, (255, 0, 0), -1)
-                viz = cv2.resize(viz, (0, 0), fx = 0.25, fy = 0.25)
-                cv2.imwrite('/home/hello-robot/alfred-autonomy/src/manipulation/scripts/cropped.png', viz)
-                self.prevtime = time.time()
-            
-            
-            # x_f = (x_new1 + x_new2)/2
-            # y_f = (y_new1 + y_new2)/2
-
-            point = self.convertPointToFrame(x_f, y_f, z_f, "base_link")
-
-            x, y, z = point.x, point.y, point.z
-
-            radius = np.sqrt(x**2 + y**2)
-            
-            confidence /= radius
-            
-            if (z > 0.7  and radius < 2.5): # to remove objects that are not in field of view
-                self.isDetected = True
-                self.objectLocArr.append((x, y, z, confidence))
-                # rospy.loginfo("Object detected at" + str((x, y, z)))
-            else:
-                self.isDetected = False
-        else:
-            self.isDetected = False
-
-    def imageCallback(self, color_img):
-        color_img = self.bridge.imgmsg_to_cv2(color_img, desired_encoding="passthrough")
-        self.color_image = np.array(color_img, dtype=np.uint8)
-        # self.depth_image = cv2.rotate(depth_img, cv2.ROTATE_90_CLOCKWISE)
-    
-
-    def depthCallback(self, depth_img):
-        depth_img = self.bridge.imgmsg_to_cv2(depth_img, desired_encoding="passthrough")
-        self.depth_image = np.array(depth_img, dtype=np.float32)
-        # self.depth_image = cv2.rotate(depth_img, cv2.ROTATE_90_CLOCKWISE)
-    
-    def reset(self):
-        self.isDetected = False
-        self.objectLocArr = []
-        self.state = State.SEARCH
-        self.requestClearObject = False
-
-    def alignObjectHorizontalTest(self, offset = 0.0):
+     
+    def alignObjectHorizontal(self, ee_pose_x = 0.0, debug_print = {}):
+        """
+        Visual servoing to align gripper with the object of interest.
+        """
         self.requestClearObject = True
-        rospy.loginfo("Waiting for clearing object")
-        if not self.clearAndWaitForNewObject(2):
+        if not self.scene_parser.clearAndWaitForNewObject(2):
             return False
 
         xerr = np.inf
         rospy.loginfo("Aligning object horizontally")
+        
         prevxerr = np.inf
-        curvel = 0.0
-        prevsettime = time.time()
-        while True:
-            x, y, z, confidence = self.objectLocArr[-1]
-            rospy.loginfo(f"{x}, {y}, {z}")
-            if self.isDetected:
-                xerr = (x) + offset
-            else:
-                rospy.logwarn("Object not detected. Using open loop motions")
-                xerr = (x - curvel * (time.time() - prevsettime)) + offset
-            rospy.loginfo("Alignment Error = " + str(xerr))
-            dxerr = (xerr - prevxerr)
-            vx = (self.kp['velx'] * xerr + self.kd['velx'] * dxerr)
-            prevxerr = xerr
-            move_to_pose(self.trajectoryClient, {
-                    'base_translate;vel' : vx,
-            }) 
-            curvel = vx
-            prevsettime = time.time()
-            
-            rospy.sleep(0.1)
-        exit()
-        move_to_pose(self.trajectoryClient, {
-            'base_translate;vel' : 0.0,
-        }) 
-        # risk
-        # if self.isDetected is False:
-        #     return False
-        return True
-    
-    def alignObjectHorizontal(self, ee_pose_x = 0.0, debug_print = None):
-        self.requestClearObject = True
-        if not self.clearAndWaitForNewObject(2):
-            return False
-
-        xerr = np.inf
-        rospy.loginfo("Aligning object horizontally")
-        prevxerr = np.inf
-        curvel = 0.0
-        prevsettime = time.time()
-        moving_avg_n = 50
+        moving_avg_n = self.moving_avg_n
         moving_avg_err = []
         mean_err = np.inf
         
-        while len(moving_avg_err) <= moving_avg_n and  mean_err > 0.008:
-            x, y, z, confidence = self.objectLocArr[-1]
-            # print(xerr)
-            if self.isDetected:
+        kp = self.pid_gains['kp']
+        kd = self.pid_gains['kd']
+        
+        while len(moving_avg_err) <= moving_avg_n and  mean_err > self.mean_err_horiz_alignment:
+            [x, y, z, conf, pred_time], isLive = self.scene_parser.get_latest_observation()
+            if isLive:
                 xerr = (x - ee_pose_x)
-                
-                
                 
             moving_avg_err.append(xerr)
             if len(moving_avg_err) > moving_avg_n:
                 moving_avg_err.pop(0)
-            # else:
-            #     rospy.logwarn("Object not detected. Using open loop motions")
-            #     xerr = (x - curvel * (time.time() - prevsettime)) + offset
-            rospy.loginfo("Alignment Error = " + str(xerr))
+            else:
+                rospy.logwarn("Object not detected. Using open loop motions")
+                xerr = (x - curvel * (time.time() - prevsettime)) - ee_pose_x
+                
             dxerr = (xerr - prevxerr)
-            vx = (self.kp['velx'] * xerr + self.kd['velx'] * dxerr)
+            vx = (kp * xerr + kd * dxerr)
             prevxerr = xerr
             
             mean_err = np.mean(np.abs(np.array(moving_avg_err)))
             debug_print["xerr"] = xerr
             debug_print["mean_err"] = mean_err
             debug_print["x, y, z"] = [x, y, z]
+            
             print(debug_print)
+
             move_to_pose(self.trajectoryClient, {
                     'base_translate;vel' : vx,
             }, asynchronous= True) 
+            
             curvel = vx
             prevsettime = time.time()
-            
             rospy.sleep(0.1)
+        
         move_to_pose(self.trajectoryClient, {
             'base_translate;vel' : 0.0,
         }, asynchronous= True) 
-        # risk
-        # if self.isDetected is False:
-        #     return False
+        
         return True
-    
-    def recoverFromFailure(self):
-        pass
 
     def main(self, objectId):
         self.objectId = objectId
-        print("Triggered Visual Servoing for object of interest:" , objectId)
+        rospy.loginfo("Triggered Visual Servoing for object of interest:" + str(objectId))
         # self.alignObjectHorizontal()
         # while not rospy.is_shutdown():
         #     rospy.sleep(0.5)
@@ -649,9 +300,14 @@ class AlignToObject:
             elif self.state == State.COMPLETE:
                 break
             rospy.sleep(0.5)
-        print("Visual Servoing Complete")
+        rospy.loginfo("Visual servoing is complete!")
         self.reset()
         return True
+    
+        
+    def reset(self):
+        self.state = State.SEARCH
+        
     
 
 if __name__ == "__main__":
