+++ src/navigation/alfred_navigation/config/navman.yaml
@@ -1,0 +1,1 @@
+n_attempts: 3
