+++ src/navigation/alfred_navigation/scripts/navigation_manager.py
@@ -14,6 +14,7 @@ from alfred_navigation.msg import NavManAction, NavManGoal, NavManResult, NavMan
 from moveback_recovery.msg import MovebackRecoveryAction, MovebackRecoveryGoal
 from utils import get_quaternion
 
+import time
 
 class NavigationManager():
 
@@ -22,7 +23,10 @@ class NavigationManager():
 
 
     def __init__(self):
-
+        
+        self.node_name = 'navigation_manager'
+        rospy.init_node(self.node_name)
+        
         self.nav_goal_x = None
         self.nav_goal_y = None
         self.nav_goal_theta = None
@@ -30,28 +34,30 @@ class NavigationManager():
         self.nav_result = None
 
         self.navman_server = actionlib.SimpleActionServer('nav_man', NavManAction, self.update_goal, False)
-
         self.movebase_client = actionlib.SimpleActionClient('move_base', MoveBaseAction)
 
         # self.moveback_client = actionlib.SimpleActionClient('moveback_recovery', MovebackRecoveryAction)
 
-        rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for move_base server...")
+        rospy.loginfo(f"[{self.node_name}]:" + "Waiting for move_base server...")
         self.movebase_client.wait_for_server()
 
         # self.moveback_client.wait_for_server()
-
-        print("Starting NavMan Server..")
+        self.n_max_attempts = rospy.get_param("/navman/n_attempts", 3)
+        self.n_attempts = 0
+        self.last_update_time = time.time()
+        
         self.navman_server.start()
+        rospy.loginfo(f"[{self.node_name}]:" + "Navigation Manager is ready!")
 
 
     def update_goal(self, goal : NavManGoal):
         
-        print("Navigation Manager received goal from Task Planner")
+        rospy.loginfo(f"[{self.node_name}]:" + "Navigation Manager received goal from Task Planner")
 
         self.nav_goal_x = goal.x
         self.nav_goal_y = goal.y
         self.nav_goal_theta = goal.theta
-
+        self.n_attempts = 0
         self.navigate_to_goal()
 
 
@@ -69,9 +75,9 @@ class NavigationManager():
         movebase_goal.target_pose.pose.orientation = quaternion
 
         self.movebase_client.send_goal(movebase_goal, feedback_cb = self.navigation_feedback)
-        print("Sending goal to move_base")
-        # print("Sending goal to move_base", self.goal_locations[locationName]['x'], self.goal_locations[locationName]['y'], self.goal_locations[locationName]['theta'])
-             
+        
+        rospy.loginfo(f"[{self.node_name}]:" + "Sending goal to move_base")
+
         wait = self.movebase_client.wait_for_result()
 
         self.check_result()
@@ -81,61 +87,61 @@ class NavigationManager():
 
         if self.movebase_client.get_state() != actionlib.GoalStatus.SUCCEEDED:
 
-            rospy.loginfo(f"[{rospy.get_name()}]:" +"Failed to reach goal")
-            
-            # cancel navigation
+            rospy.loginfo(f"[{self.node_name}]:" +"Failed to reach goal. Debugging...")
             self.movebase_client.cancel_goal()
-
-            self.nav_result = False
-
             self.debug_nav_failure()
-
-
         else:
-            self.nav_result = True
-            self.send_navman_result()
+            self.send_navman_result(success = True)
 
         
     
     def debug_nav_failure(self):
+        # TODO - in future, add specific conditions where things can be fixed.
+        # For now, just try again until success.
 
-            # print("Sending moveback goal!")
-            # moveback_goal = MovebackRecoveryGoal()
-            # moveback_goal.execute = True
-            # self.moveback_client.send_goal(moveback_goal)
+        # print("Sending moveback goal!")
+        # moveback_goal = MovebackRecoveryGoal()
+        # moveback_goal.execute = True
+        # self.moveback_client.send_goal(moveback_goal)
 
-            # print("Successfully sent Moveback goal!")
+        # print("Successfully sent Moveback goal!")
 
-            print("Sending goal again!")
-    
+        if self.n_attempts < self.n_max_attempts:
+            self.n_attempts += 1
+            rospy.loginfo(f"[{self.node_name}]:" + f"Trying again: {self.n_max_attempts - self.n_max_attempts} left")
             self.navigate_to_goal()
-
+        else:
+            self.send_navman_result(success = False)
         
         
-    def send_navman_result(self):
+    def send_navman_result(self, success):
 
         navman_result = NavManResult()
 
-        if self.nav_result:
+        if success:
             navman_result.success = True
             self.navman_server.set_succeeded(navman_result)
-            print("We have successfully reached the goal!!")
+            rospy.loginfo(f"[{self.node_name}]:" + f"Goal reached successfully.")
+        else:
+            navman_result.success = False
+            self.navman_server.set_aborted(navman_result)
+            rospy.loginfo(f"[{self.node_name}]:" + f"Goal could not be reached.")
 
 
     def navigation_feedback(self, msg : MoveBaseFeedback):
         self.position = msg.base_position.pose.position
         self.orientation = msg.base_position.pose.orientation
         
-        # if time.time() - self.last_update_time > 1:
-        #     self.last_update_time = time.time()
-        rospy.logdebug(f"[{rospy.get_name()}]:" +"Current position: {}".format(self.position))
-        rospy.logdebug(f"[{rospy.get_name()}]:" +"Current orientation: {}".format(self.orientation))
-
+        if time.time() - self.last_update_time > 1:
+            self.last_update_time = time.time()
+            rospy.loginfo(f"[{self.node_name}]:" +"Current position: {}".format(self.position))
+            rospy.loginfo(f"[{self.node_name}]:" +"Current orientation: {}".format(self.orientation))
+        
 
 
 if __name__ == '__main__':
 
-    rospy.init_node('navigation_manager')
+    
     node = NavigationManager()
 
     rospy.spin()
