+++ src/planning/task_planner/scripts/state_manager.py
@@ -170,20 +170,6 @@ class GlobalStates(Enum):
     CALL = 8
     
 
-# class ObjectOfInterest(Enum): 
-#     # FILL WITH CLASS LABELS FOR THE OBJECT DETECTION PIPELINE
-#     NONE = -1
-#     SODA = 1
-#     USER = 0
-#     BOTTLE = 39
-#     BOX = 2
-#     GLASS = 3
-#     TABLE = 60
-#     REMOTE = 65
-#     APPLE = 47
-#     BANANA = 46
-    
-
 class LocationOfInterest(Enum):
     # FILL WITH LOCATION LABELS FROM GROUNDING PIPELINE
     HOME = -1
