+++ src/planning/task_planner/scripts/task_planner.py
@@ -24,85 +24,63 @@ import rospkg
 import rospy
 from std_srvs.srv import Trigger, TriggerResponse
 from control_msgs.msg import FollowJointTrajectoryAction
-from firebase_node import FirebaseNode
-import threading
-import concurrent.futures
 
-# from alfred_msgs.msg import Speech, SpeechTrigger
-# from alfred_msgs.srv import GlobalTask, GlobalTaskResponse, VerbalResponse, VerbalResponseRequest, GlobalTaskRequest
+from alfred_msgs.msg import Speech, SpeechTrigger
+from alfred_msgs.srv import GlobalTask, GlobalTaskResponse, VerbalResponse, VerbalResponseRequest, GlobalTaskRequest
 from move_base_msgs.msg import MoveBaseAction, MoveBaseGoal, MoveBaseActionResult, MoveBaseResult
-# from manipulation.msg import TriggerAction, TriggerFeedback, TriggerResult, TriggerGoal
+from manipulation.msg import TriggerAction, TriggerFeedback, TriggerResult, TriggerGoal
 from alfred_navigation.msg import NavManAction, NavManGoal
-
-
-# import sys
-# sys.path.append("/home/hello-robot/ws/src/navigation/alfred_navigation/scripts")
-# from navigation_manager import NavigationManager
-
-
-from state_manager import BotStateManager, Emotions, LocationOfInterest, GlobalStates, VerbalResponseStates, OperationModes
+from state_manager import BotStateManager, Emotions, LocationOfInterest, GlobalStates, ObjectOfInterest, VerbalResponseStates, OperationModes
 import os
 import json
 from enum import Enum
-import pyrebase
-import time
 
 from utils import get_quaternion
 
 class TaskPlanner:
     def __init__(self):
-        self.trigger=0
-        # self.q=[]
-        self.q = []
-        self.last_task_success=0
         rospy.init_node('task_planner')
+
+        self.wakeword_topic_name = rospy.get_param("wakeword_topic_name",
+                             "/interface/wakeword_detector/wakeword")
+        
         rospack = rospkg.RosPack()
         base_dir = rospack.get_path('task_planner')
         locations_file = rospy.get_param("locations_file", "config/locations.json")
         locations_path = os.path.join(base_dir, locations_file)
+
         self.goal_locations = json.load(open(locations_path))
-        
-        self.executor = concurrent.futures.ThreadPoolExecutor(max_workers=1)
-
-        firebase_secrets_path = os.path.expanduser("~/firebasesecrets.json")
-        if not os.path.isfile(firebase_secrets_path):
-            raise FileNotFoundError("Firebase secrets file not found")
-        
-        with open(firebase_secrets_path, 'r') as f:
-            config = json.load(f)
-
-        # with open(os.path.expanduser("~/alfred-autonomy/src/planning/task_planner/config/firebase_schema.json")) as f:
-        #     self.state_dict = json.load(f)
-
-        self.firebase = pyrebase.initialize_app(config)
-        self.db = self.firebase.database()
-        # self.db.remove("")
-        # self.state_dict["button_callback"] = 0
-        # self.state_dict["button_callback2"] = 0
-        # self.state_dict["button_callback3"] = 0
-        # self.state_dict["button_callback4"] = 0
-
-        # self.db.set(self.state_dict) # fill
- 
-        self.last_update = time.time()
-
-        self.bot_state = BotStateManager(self.stateUpdateCallback)
-
-        self.nav_test = True
-
-        rospack = rospkg.RosPack()
         self.startManipService = rospy.ServiceProxy('/switch_to_manipulation_mode', Trigger)
         self.startNavService = rospy.ServiceProxy('/switch_to_navigation_mode', Trigger)
 
-        # self.navigation_client = actionlib.SimpleActionClient('move_base', MoveBaseAction)
-        #self.commandReceived = rospy.Service('/robot_task_command', GlobalTask, self.command_callback)
+        self.navigation_client = actionlib.SimpleActionClient('move_base', MoveBaseAction)
+        self.manipulation_client = actionlib.SimpleActionClient('manipulation_fsm', TriggerAction)
+
+        self.verbal_response_service_name = rospy.get_param("verbal_response_service_name",
+                                "/interface/response_generator/verbal_response_service")
+
+        self.startedListeningService = rospy.Service('/startedListening', Trigger, self.wakeWordTriggered)
+        self.commandReceived = rospy.Service('/robot_task_command', GlobalTask, self.command_callback)
         
+        self.heightOfObject = 0.84 # change this to a local variable later on 
+ 
         self.stow_robot_service = rospy.ServiceProxy('/stow_robot', Trigger)
         rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for stow robot service...")
-        # self.stow_robot_service.wait_for_service()
+        self.stow_robot_service.wait_for_service()
 
-        # rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for move_base server...")
-        # self.navigation_client.wait_for_server()
+        self.verbal_response_service = rospy.ServiceProxy(self.verbal_response_service_name, VerbalResponse)
+        rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for verbal response service...")
+        self.verbal_response_service.wait_for_service()
+
+        rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for move_base server...")
+        self.navigation_client.wait_for_server()
+
+        rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for manipulation_fsm server...")
+        self.manipulation_client.wait_for_server()
+
+        self.bot_state = BotStateManager(self.stateUpdateCallback)
+        rospy.loginfo(f"[{rospy.get_name()}]:" + "Bot state manager initialized")
+
 
         self.trajectoryClient = actionlib.SimpleActionClient('alfred_controller/follow_joint_trajectory', FollowJointTrajectoryAction)
         rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for driver..")
@@ -111,18 +89,10 @@ class TaskPlanner:
 
         self.task_requested = False
         self.time_since_last_command = rospy.Time.now()
+        self.startManipService = rospy.ServiceProxy('/switch_to_manipulation_mode', Trigger)
         self.startManipService.wait_for_service()
-
         self.isManipMode = False
         self.runningTeleop = False
-
-        self.mappings = {
-            # 'fetch_water' : [LocationOfInterest.LIVING_ROOM, ObjectOfInterest.BOTTLE],
-            # 'fetch_apple' : [LocationOfInterest.NET, ObjectOfInterest.APPLE],
-            # 'fetch_remote' : [LocationOfInterest.LIVING_ROOM, ObjectOfInterest.REMOTE],
-            # 'fetch_banana' : [LocationOfInterest.LIVING_ROOM, ObjectOfInterest.BANANA],
-            # 'fetch_soda' : [LocationOfInterest.LIVING_ROOM, ObjectOfInterest.SODA],
-        }
 
         self.class_list = rospy.get_param("/object_detection/class_list")
         
@@ -139,11 +109,15 @@ class TaskPlanner:
         rospy.loginfo(f"[{rospy.get_name()}]:" + "Waiting for Navigation Manager server...")
         self.navigation_client.wait_for_server()
 
-        self.navigate_to_location_manual()
-
+        rospy.loginfo(f"[{rospy.get_name()}]:" + "Node Ready.")
         # create a timed callback to check if we have, Obj been idle for too long
 
-    def stateUpdateCallback(self):        
+    def idleCallback(self, event):
+        if not self.bot_state.isAutonomous() and self.bot_state.isExtendedCommunicationLossTeleop():
+            self.bot_state.currentGlobalState = GlobalStates.IDLE
+
+    def stateUpdateCallback(self):
+        
         if not self.bot_state.isAutonomous():
             if not self.isManipMode :
                 self.isManipMode = True
@@ -151,8 +125,6 @@ class TaskPlanner:
             
             if not self.runningTeleop:
                 self.maintainvelocities()
-    def delay(self):
-        time.sleep(5)
 
     def maintainvelocities(self):
         self.runningTeleop = True
@@ -172,173 +144,84 @@ class TaskPlanner:
             if commands['manipulator']['gripper_open']:
                 gripperpos = 100
             move_to_pose(self.trajectoryClient, 
-                {
-                    'stretch_gripper;to' : gripperpos,
-                    'wrist_yaw;to' : commands['manipulator']['yaw_position'],
-                    'lift;vel' : commands['manipulator']['vel_lift'],
-                    'arm;vel' : commands['manipulator']['vel_extend']
-                },
+                {'stretch_gripper;to' : gripperpos},
+            )
+            move_to_pose(self.trajectoryClient, 
+                {'wrist_yaw;to' : commands['manipulator']['yaw_position']},
+            )
+            move_to_pose(self.trajectoryClient, 
+                {'lift;vel' : commands['manipulator']['vel_lift']},
+            )
+            move_to_pose(self.trajectoryClient, 
+                {'arm;vel' : commands['manipulator']['vel_extend']},
             )
             
             rospy.sleep(0.1)
-    def main(self):
-        task_thread = threading.Thread(target=task_planner.execute_task_temp)
-        task_thread2 = threading.Thread(target=task_planner.updateButton)
-        # task_thread.start()
-        # task_thread2.start()
 
-        # task_thread.join()
-        # task_thread2.join()
+        self.runningTeleop = False
+        print("teleop stopped")
 
-    def execute_task_temp(self):
-        
-        while True:
-            try:
-                if len(self.q) != 0 or self.nav_test == True:
-                    # remote_id, task_to_execute = self.q.pop(0)
-                    tasks_to_execute = self.q.pop(0)
-                    # execute tasks here.
-                    rospy.loginfo(f"Executing task {tasks_to_execute} for room {1}")
-                    self.navigate_to_location(self.navigationGoal, tasks_to_execute)
-                    rospy.sleep(5)
-                    rospy.loginfo("Task complete...")
-                    self.nav_test = False
-            except KeyboardInterrupt:
-                break
+    def executeTask(self):
 
+        if not self.bot_state.isAutonomous():
+            rospy.loginfo("Received a command, but not in autonomous mode, so ignoring. Ask to relenquish control.")
+            return
 
-        # if(len(self.q)<4):
-        #     if(len(self.q)!=4):
-        #         self.q.append(0)
-        #         if(len(self.q)!=4):
-        #             self.q.append(0)
-        # if((self.q)and(len(self.q)>3)):
-        #     # task_thread = threading.Thread(target=self.executeTask, args=(0,self.q[0]))
-        #     # self.executeTask(0, self.q[0])
-        #     self.trigger=1
-        #     if(self.q[0]==1):
-        #         rospy.loginfo("Jake requested for waterr.")
-        #         self.q.popleft()
-        #         # self.last_task_success=self.q[0]
-        #         self.delay()
-        #     elif(self.q[0]==2):
-        #         rospy.loginfo("Jake wants to talk to Donna.")
-        #         self.q.popleft()
-        #         time.sleep(5)
-        #     elif(self.q[0]==3):
-        #         rospy.loginfo("Steve wants water.")
-        #         self.q.popleft()
-
-        #         # self.last_task_success=self.q[0]
-        #         time.sleep(5)
-        #     elif(self.q[0]==4):
-        #         rospy.loginfo("Steve wants to talk to Harry.")
-        #         self.q.popleft()
-
-        #             # self.last_task_success=self.q[0]
-        #         time.sleep(5)
-        #     elif(self.q[0]==0):
-                
-        #         rospy.loginfo(self.q[0])
-        #         rospy.loginfo("No tasks yet.")
-        #         self.q.popleft()
-                        # task_thread.start()
-                        # task_thread = self.executor.submit(self.executeTask, 0, q[0])
-                        
-                # task_thread.start()
-                # if self.last_task_success == self.q[0]:
-                #     self.q.pop()
-                #     rospy.loginfo(self.q)
-                #     # rospy.sleep(5)
-                #     self.last_task_success=0
-
-       
-                # self.startNavService()
-                # success = self.navigate_to_location(LocationOfInterest.TABLE)
-                # self.bot_state.update_operation_mode(OperationModes.TELEOPERATION)
-                # success = self.navigate_to_location(LocationOfInterest.HOME)
-
-    def executeTask(self,i,j):
-
-        # self.stow_robot_service()
-        # self.startNavService()
-        k=j
-        if(k==1):
-            rospy.loginfo("Jake requested for water.")
-            self.last_task_success=k
-            rospy.sleep(5)
-        elif(k==2):
-            rospy.loginfo("Jake wants to talk to Donna.")
-            self.last_task_success=k
-            rospy.sleep(5)
-        elif(k==3):
-            rospy.loginfo("Steve wants water.")
-            self.last_task_success=k
-            rospy.sleep(5)
-        elif(k==4):
-            rospy.loginfo("Steve wants to talk to Harry.")
-            self.last_task_success=k
-            rospy.sleep(5)
-        elif(k==0):
-            rospy.loginfo(k)
-            rospy.loginfo("No tasks yet.")
-        navSuccess = self.navigate_to_location(self.navigationGoal,j)
-        return self.last_task_success
-        
-    # def jake(self):
-    #     self.startNavService()
-    #     navSuccess = self.navigate_to_location(self.navigationGoal)
-
-    def navigate_to_location(self, location : Enum, k):
-        locationName = location.name
-        goal = NavManGoal()
-        rospy.loginfo(f"[{rospy.get_name()}]:" +"Executing task. Going to {}".format(locationName))
-        # send goal
-        goal.x = -23.77337646484375 #elevator
-        goal.y = -23.77337646484375 #elevator
-        goal.theta = 240
-
-        # self.navigation_manager.navigate_to_goal(goal.target_pose.pose.position.x, goal.target_pose.pose.position.y, 240)
+        rospy.loginfo("Task has been requested")
+        rospy.loginfo("Sending verbal response to acknowledge task reception.")
+        self.process_verbal_response(VerbalResponseStates.ON_IT)
 
         self.stow_robot_service()
         self.startNavService()
         rospy.sleep(0.5)
-        rospy.loginfo("Going to {}, to manipulate {}".format(self.navigationGoal.name, self.class_list[self.objectOfInterest]))
+        rospy.loginfo("Going to {}, to manipulate {}".format(self.navigationGoal.name, self.objectOfInterest.name))
         
         navSuccess = self.navigate_to_location(self.navigationGoal)
         if not navSuccess:
             rospy.loginfo("Failed to navigate to table, adding intermediate waypoints")
             navSuccess = self.navigate_to_location(self.navigationGoal)
-        # goal.target_pose.header.frame_id = "map"
-        # quaternion = get_quaternion(240)
-        # goal.target_pose.pose.orientation = quaternion
-        self.navigation_client.send_goal(goal, feedback_cb = self.bot_state.navigation_feedback)
-        # print("Sending goal to move_base", self.goal_locations[locationName]['x'], self.goal_locations[locationName]['y'], self.goal_locations[locationName]['theta'])
         
-             
-        # wait = self.navigation_client.wait_for_result()
+        self.startManipService()
+        rospy.sleep(0.5)
 
-        # if self.navigation_client.get_state() != actionlib.GoalStatus.SUCCEEDED:
-        #     rospy.loginfo(f"[{rospy.get_name()}]:" +"Failed to reach Table")
-        #     # cancel navigation
-        #     self.navigation_client.cancel_goal()
-        #     return False
+        self.bot_state.currentGlobalState = GlobalStates.MANIPULATION
+        rospy.loginfo("Attempting to find and manipulate objects.")
+        manipulationSuccess = self.manipulate_object(self.objectOfInterest, isPick = True)
+        if not manipulationSuccess:
+            rospy.loginfo("Manipulation failed. Cannot find object, returning to user")
         
-        # rospy.loginfo(f"[{rospy.get_name()}]:" +"Reached Table")
-        # self.bot_state.update_emotion(Emotions.HAPPY)
-        # self.bot_state.update_state()
-        # self.bot_state.currentGlobalState = GlobalStates.REACHED_GOAL
-        # rospy.loginfo(f"[{rospy.get_name()}]:" +"Reached Table")
-        return True
-    
+        self.stow_robot_service()
+        self.startNavService()
+        rospy.sleep(0.5)                                                        
+        rospy.loginfo(f"Going to table. Manipulation success = {manipulationSuccess}")
+        success = self.navigate_to_location(LocationOfInterest.TABLE)
+        if not success:
+            rospy.loginfo("Failed to navigate to table, adding intermediate waypoints")
+            self.navigate_to_location(LocationOfInterest.TABLE)
+            # add waypoints
 
-    def manipulate_object(self, object, isPick = True):
+        self.bot_state.currentGlobalState = GlobalStates.MANIPULATION
+        self.startManipService()
+
+        if manipulationSuccess:
+            rospy.loginfo("Requesting to place the object")
+            self.manipulate_object(self.objectOfInterest, isPick = False)
+            self.process_verbal_response(VerbalResponseStates.HERE_YOU_GO)
+        else:
+            rospy.loginfo("Requesting verbal command.")
+            self.process_verbal_response(VerbalResponseStates.SORRY)
+
+        self.bot_state.currentGlobalState = GlobalStates.IDLE
+        self.task_requested = False
+        self.time_since_last_command = rospy.Time.now()
+
+
+    def manipulate_object(self, object : Enum, isPick = True):
         rospy.loginfo("Starting manipulation, going to perform object" + (" picking" if isPick else " placing"))
-        rospy.loginfo("Manipulating object")
-        
+        rospy.loginfo("Manipulating {}".format(object.name))
         self.bot_state.currentGlobalState = GlobalStates.MANIPULATION
         # send goal
-        goalObject = object
+        goalObject = object.value
         if not isPick:
             goalObject = 60 #table
 
@@ -383,111 +266,62 @@ class TaskPlanner:
             rospy.loginfo("Updating operation mode to autonomous")
             self.bot_state.update_operation_mode(OperationModes.AUTONOMOUS)
         elif msg.primitive != 'do_nothing':
-            self.navigationGoal, self.objectOfInterest = self.mappings[(msg.primitive)]
+            self.navigationGoal, self.objectOfInterest = self.mappings[msg.primitive]
             self.executeTask()
 
         return GlobalTaskResponse(success = True)
-        
-
-    def navigate_to_location_manual(self):
-        # locationName = location.name
+    
+    
+    def navigate_to_location_navman(self, location : Enum):
+        locationName = location.name
         goal = NavManGoal()
-        # rospy.loginfo(f"[{rospy.get_name()}]:" +"Executing task. Going to {}".format(locationName))
-        # send goal
-        # goal.x = -3.8682618141174316 #elevator
-        # goal.y = -2.505342483520508 #elevator
-        # goal.theta = 240
-
-        # goal.x = -4.1436285972595215 
-        # goal.y = -4.530055522918701 
-        # goal.theta = 240
-
-        # near AIMS fridge
-        goal.x = -2.615337371826172
-        goal.y = -7.306527137756348
-        goal.theta = 0
-
-
-        # # near AIMS Baxter
-        # goal.x = -1.9814379215240479
-        # goal.y = -0.4820127487182617
-        # goal.theta = 0
+        goal.x = self.goal_locations[locationName]['x']
+        goal.y = self.goal_locations[locationName]['y']
+        goal.theta = self.goal_locations[locationName]['theta']
         
         self.navigation_client.send_goal(goal, feedback_cb = self.bot_state.navigation_feedback)
+        wait = self.navigation_client.wait_for_result()
 
-        print("Task Planner has sent goal to Navigation Manager")
+        if self.navigation_client.get_state() != actionlib.GoalStatus.SUCCEEDED:
+            rospy.loginfo(f"[{rospy.get_name()}]:" +"Failed to reach {}".format(locationName))
+            self.navigation_client.cancel_goal()
+            return False
+    
+        return True
 
+    def navigate_to_location(self, location : Enum):
+        locationName = location.name
+        rospy.loginfo(f"[{rospy.get_name()}]:" +"Executing task. Going to {}".format(locationName))
+        # send goal
+        print("Sending goal to move_base", self.goal_locations[locationName]['x'], self.goal_locations[locationName]['y'], self.goal_locations[locationName]['theta'])
+        goal = MoveBaseGoal()
+        goal.target_pose.header.frame_id = "map"
+        goal.target_pose.pose.position.x = self.goal_locations[locationName]['x']
+        goal.target_pose.pose.position.y = self.goal_locations[locationName]['y']
+        quaternion = get_quaternion(self.goal_locations[locationName]['theta'])
+        goal.target_pose.pose.orientation = quaternion
+        self.navigation_client.send_goal(goal, feedback_cb = self.bot_state.navigation_feedback)
+        wait = self.navigation_client.wait_for_result()
+
+        if self.navigation_client.get_state() != actionlib.GoalStatus.SUCCEEDED:
+            rospy.loginfo(f"[{rospy.get_name()}]:" +"Failed to reach {}".format(locationName))
+            # cancel navigation
+            self.navigation_client.cancel_goal()
+            return False
+        
+        rospy.loginfo(f"[{rospy.get_name()}]:" +"Reached {}".format(locationName))
+        self.bot_state.update_emotion(Emotions.HAPPY)
+        self.bot_state.update_state()
+        self.bot_state.currentGlobalState = GlobalStates.REACHED_GOAL
+        rospy.loginfo(f"[{rospy.get_name()}]:" +"Reached {}".format(locationName))
         return True
 
 
-
-    def updateButton(self):
-        while True:
-            try:
-                rospy.loginfo(self.q)
-                # remote_callback_0 : 0
-                # remote_callback_1 : 2
-                # remote_callback_2 : 3
-                # remote_callback_3 : 4
-
-                # for remoteid in range(10):
-                #     remoteval = self.db.child(f'remote_callback_{remoteid}').get().val()
-
-                #     if remoteval != 0:
-                #         self.q.append((remoteid, remoteval))
-
-                button_callback_value = self.db.child("button_callback").get().val()
-                button_callback_value2 = self.db.child("button_callback2").get().val()
-                button_callback_value3 = self.db.child("button_callback3").get().val()
-                button_callback_value4 = self.db.child("button_callback4").get().val()
-                button_callback_value5 = self.db.child("button_callback5").get().val()
-                button_callback_value6 = self.db.child("button_callback6").get().val()
-                button_callback_value7 = self.db.child("button_callback7").get().val()
-                button_callback_value8 = self.db.child("button_callback8").get().val()
-                if(button_callback_value==1):
-                    if 1 not in self.q:
-                        self.q.append(1)
-                        self.db.child("button_callback").set(0)
-
-                elif(button_callback_value2==1):
-                    if 2 not in self.q:
-                        self.q.append(2)
-                        self.db.child("button_callback2").set(0)
-                elif(button_callback_value3==1):
-                    if 3 not in self.q:
-                        self.q.append(3)
-                        self.db.child("button_callback3").set(0)
-                elif(button_callback_value4==1):
-                    if 4 not in self.q:
-                        self.q.append(4)   
-                        self.db.child("button_callback4").set(0)
-                elif(button_callback_value5==1):
-                    if 5 not in self.q:
-                        self.q.append(5)   
-                        self.db.child("button_callback5").set(0)
-                elif(button_callback_value6==1):
-                    if 6 not in self.q:
-                        self.q.append(6)   
-                        self.db.child("button_callback6").set(0)
-                elif(button_callback_value7==1):
-                    if 7 not in self.q:
-                        self.q.append(7)   
-                        self.db.child("button_callback7").set(0)
-                elif(button_callback_value8==1):
-                    if 8 not in self.q:
-                        self.q.append(8)   
-                        self.db.child("button_callback8").set(0)
-            except KeyboardInterrupt:
-                break
-
 if __name__ == "__main__":
-
     task_planner = TaskPlanner()
-    
 
-    # rospy.sleep(2)
-    try:    
-        task_planner.main()
+    try:
+        rospy.spin()
     except KeyboardInterrupt:
         print("Shutting down")
